"preparing for release..."

+++ dev-scripts/cleanup.sh
@@ -3,6 +3,11 @@
 
 dir=$1
 
+echo "Cleaning $dir..."
+
+[[ -z "$dir" ]] && { echo "Error: First argument should be a path to user home dir"; exit 1; }
+
 find $dir -size 0 -delete -print
 find $dir -empty -type d -delete -print
-
+rm -rf $dir/etc/
+rm -rf $dir/run/
